pyproject.toml
[tool.poetry]
name = "good_morning_project"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
pandas = "^2.0.3"
streamlit = "^1.22.0"
plost = "^0.2.5"
matplotlib = "^3.7.1"
scipy = "^1.10.1"
mysql-connector-python = "^8.0.33"
google-api-python-client = "^2.83.0"
google-auth-oauthlib = "^1.0.0"
seaborn = "^0.12.2"
xlsxwriter = "^3.0.9"
plotly = "^5.15.0"
openpyxl = "^3.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
